bonanza: increase effect speed progression
Addresses #5

The timestep variable is responsible for controlling the speed of the
animation. 2.0 is too low of a timestep value since this makes it at
least four times slower than other available animations.

So why 8.0 for the timestep value?

I was comparing the bonanza animation speed to the blinds animation
speed. The new timestep value puts its speed at a similar completion
rate to blinds.